refactor: debug module structure
